+++ src/types.rs
@@ -87,7 +87,7 @@ pub struct ForwardRequest {
     /// paymentToken for Gelato Executors
     pub fee_token: H160,
     ///Type identifier for Gelato's payment. Can be 1, 2 or 3.
-    pub payment_type: usize, // 1 = gas tank
+    pub payment_type: PaymentType, // 1 = gas tank
     /// Maximum fee sponsor is willing to pay Gelato Executors
     pub max_fee: U64,
     /// Gas limit