+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Unreleased
 
+- fix: use PaymentType type for payment types
+
 # v0.1.0-alpha
 
 - adds bindings getting estimated relayer fee