+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 # Unreleased
 
+- refactor: Rename the dispatch-ready forward req to `Signed___`.
+- feat: Move the unsigned forward req into this crate
+- refactor: Clean up typing
+- refactor: Start moving towards `Address` over `H160`
 - fix: use PaymentType type for payment types
 
 # v0.1.0-alpha