+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Unreleased
 
+- feature: support ForwardCall (type 0 payments)
 - refactor: move all rpc requests into modules
 - refactor: Rename the dispatch-ready forward req to `Signed___`.
 - feat: Move the unsigned forward req into this crate